Added a JSON link along with some other changes
I added a bit more in the script so you can search up "Block Party" (case sensitive) and such and it will still return. I plan to add more keywords to make it recognize even "block party" and stuff like that.

Also created a lone hidePanel script opposed to using the PCN one since I added a new div that should be hidden until the button is clicked, and hidden again if the Hide Panel button is clicked.